move extractInlineImagesToAttachments() to ComposeSendBtnModule
